Move engine run test and fix examples
