+++ tests/test_template_format.py
@@ -3,8 +3,7 @@ from pathlib import Path
 
 from prompti.loader import FileSystemLoader
 from prompti.template import PromptTemplate, Variant
-from prompti.model_client import ModelClient, ModelConfig, RunParams
-from prompti.message import Message
+from prompti.model_client import ModelConfig
 
 
 @pytest.mark.asyncio
@@ -95,35 +94,4 @@ def test_complex_jinja_multi_message():
     msgs, _ = template.format({"tasks": tasks}, variant="base")
     content = msgs[0].content
     assert "Fix" in content and "Doc" in content
-
-
-@pytest.mark.asyncio
-async def test_template_run_uses_model_cfg():
-    class DummyClient(ModelClient):
-        provider = "dummy"
-
-        async def _run(self, params: RunParams):
-            yield Message(role="assistant", kind="text", content="ok")
-
-    tmpl_cfg = ModelConfig(provider="dummy", model="x")
-    client = DummyClient(ModelConfig(provider="dummy", model="y"))
-    template = PromptTemplate(
-        name="x",
-        description="",
-        version="1",
-        variants={
-            "base": Variant(
-                contains=[],
-                model_config=tmpl_cfg,
-                messages=[{"role": "user", "parts": []}],
-            )
-        },
-    )
-
-    out = [
-        m
-        async for m in template.run({}, client=client, variant="base", stream=False)
-    ]
-    assert out[0].content == "ok"
-    assert client.cfg == tmpl_cfg
